[windows] initial revision.
Adds windows support (w/o wmi) for collecting process stats.
Uses specific branch of datadog/gopsutil
  - Note: hand-edited glide.lock to get correct version.  This is bad.
Adds prelim windows service processing.
Adds version resource on windows
Adds message file for reporting to event viewer
