+++ .gitignore
@@ -1,2 +1,10 @@
 dd-process-agent
 vendor/
+
+.vscode
+*.exe
+*.syso
+agent/windows_resources/process-agent-msg.rc
+agent/windows_resources/*.bin
+# file generated at build time by message compiler
+process-agent-msg.h
