+++ agent/windows_resources/process-agent-msg.mc
@@ -1,0 +1,85 @@
+;// Header
+MessageIdTypedef=DWORD
+
+LanguageNames=(
+    English=0x409:MSG00409
+)
+
+
+;// Messages
+MessageId=1
+SymbolicName=MSG_WARN_REGCONFIG_FAILED
+Severity=Warning
+Language=English
+Failed to import config items from registry.  The error was %1
+.
+
+MessageId=2
+SymbolicName=MSG_WARN_CONFIGUPGRADE_FAILED
+Severity=Warning
+Language=English
+Failed to upgrade configuration.  The error was %1.
+.
+
+MessageId=3
+SymbolicName=MSG_SERVICE_STARTED
+Severity=Informational
+Language=English
+The %1 service has started.
+.
+
+MessageId=4
+SymbolicName=MSG_SERVICE_STOPPED
+Severity=Informational
+Language=English
+The %1 service has stopped.
+.
+
+MessageId=5
+SymbolicName=MSG_UNKNOWN_CONTROL_REQUEST
+Severity=Warning
+Language=English
+Unexpected control request %1
+.
+
+MessageId=6
+SymbolicName=MSG_SERVICE_STOPPING
+Severity=Informational
+Language=English
+Received stop command, shutting down
+.
+
+MessageId=7
+SymbolicName=MSG_SERVICE_STARTED
+Severity=Informational
+Language=English
+starting the %1 service.
+.
+
+MessageId=8
+SymbolicName=MSG_SERVICE_FAILED
+Severity=Error
+Language=English
+The Service failed: %1
+.
+
+MessageId=9
+SymbolicName=MSG_SERVICE_FAILED
+Severity=Warning
+Language=English
+The Service failed.  %1
+.
+
+MessageId=10
+SymbolicName=MSG_UNEXPECTED_CONTROL_REQUEST
+Severity=Error
+Language=English
+The Service failed: %1
+.
+
+MessageId=11
+SymbolicName=MSG_RECEIVED_STOP_COMMAND
+Severity=Informational
+Language=English
+The service %1 received the stop command, shutting down.
+.
