+++ agent/windows_resources/process-agent.rc
@@ -1,0 +1,54 @@
+#include "version.h"
+#define RT_MANIFEST 24
+#define APP_MANIFEST 1
+
+
+//APP_MANIFEST RT_MANIFEST agent.exe.manifest
+
+#define MAIN_ICON 1
+#define ICON_16_16 2
+#define ICON_32_32 3
+
+MAIN_ICON ICON "project.ico"
+ICON_16_16 ICON "project_16x16.ico"
+ICON_32_32 ICON "project_32x32.ico"
+
+#define VS_VERSION_INFO 1
+VS_VERSION_INFO VERSIONINFO
+ FILEVERSION RC_FILE_VERSION
+ PRODUCTVERSION RC_FILE_VERSION
+ FILEFLAGSMASK 0x3fL
+#ifdef _DEBUG
+ FILEFLAGS 0x1L
+#else
+ FILEFLAGS 0x0L
+#endif
+ FILEOS 0x40004L
+ FILETYPE 0x0L
+ FILESUBTYPE 0x0L
+BEGIN
+    BLOCK "StringFileInfo"
+    BEGIN
+        BLOCK "040904b0"
+        BEGIN
+            VALUE "CompanyName", "Datadog, Inc."
+            VALUE "FileDescription", "Datadog Trace Agent"
+            VALUE "FileVersion", FILE_VERSION_STRING
+            VALUE "InternalName", "trace-agent"
+            VALUE "LegalCopyright", "Copyright (C) 2017"
+            VALUE "OriginalFilename", "trace-agent.exe"
+            VALUE "ProductName", "Datadog Trace Agent"
+            VALUE "ProductVersion", FILE_VERSION_STRING
+        END
+    END
+    BLOCK "VarFileInfo"
+    BEGIN
+        VALUE "Translation", 0x409, 1200
+    END
+END
+
+// Country: United States
+// Language: English
+#pragma code_page(437)
+LANGUAGE 0x9, 0x1
+1 MESSAGETABLE "MSG00409.bin"
