+++ agent/windows_resources/version.h
@@ -1,0 +1,7 @@
+
+#define RC_FILE_VERSION MAJ_VER,MIN_VER,PATCH_VER,0
+
+#define STRINGIFY(x) #x
+#define TO_STRING(x) STRINGIFY(x)
+
+#define FILE_VERSION_STRING TO_STRING(MAJ_VER.MIN_VER.PATCH_VER.0)
