+++ checks/process.go
@@ -1,9 +1,6 @@
 package checks
 
 import (
-	"os/user"
-	"runtime"
-	"strconv"
 	"time"
 
 	"github.com/DataDog/gopsutil/cpu"
@@ -221,26 +218,6 @@ func formatIO(fp *process.FilledProcess, lastIO *process.IOCountersStat, before 
 	}
 }
 
-func formatUser(fp *process.FilledProcess) *model.ProcessUser {
-	var username string
-	var uid, gid int32
-	if len(fp.Uids) > 0 {
-		u, err := user.LookupId(strconv.Itoa(int(fp.Uids[0])))
-		if err == nil {
-			username = u.Username
-		}
-		uid = int32(fp.Uids[0])
-	}
-	if len(fp.Gids) > 0 {
-		gid = int32(fp.Gids[0])
-	}
-
-	return &model.ProcessUser{
-		Name: username,
-		Uid:  uid,
-		Gid:  gid,
-	}
-}
 
 func formatMemory(fp *process.FilledProcess) *model.MemoryStat {
 	ms := &model.MemoryStat{
@@ -259,21 +236,6 @@ func formatMemory(fp *process.FilledProcess) *model.MemoryStat {
 	return ms
 }
 
-func formatCPU(fp *process.FilledProcess, t2, t1, syst2, syst1 cpu.TimesStat) *model.CPUStat {
-	numCPU := float64(runtime.NumCPU())
-	deltaSys := syst2.Total() - syst1.Total()
-	return &model.CPUStat{
-		LastCpu:    t2.CPU,
-		TotalPct:   calculatePct((t2.User-t1.User)+(t2.System-t1.System), deltaSys, numCPU),
-		UserPct:    calculatePct(t2.User-t1.User, deltaSys, numCPU),
-		SystemPct:  calculatePct(t2.System-t1.System, deltaSys, numCPU),
-		NumThreads: fp.NumThreads,
-		Cpus:       []*model.SingleCPUStat{},
-		Nice:       fp.Nice,
-		UserTime:   int64(t2.User),
-		SystemTime: int64(t2.System),
-	}
-}
 
 func calculatePct(deltaProc, deltaTime, numCPU float64) float32 {
 	if deltaTime == 0 {
