+++ glide.lock
@@ -42,7 +42,7 @@ imports:
   - processes
   - processes/gops
 - name: github.com/DataDog/gopsutil
-  version: df069044fce2744779f4ba1fb60c12938b4f2951
+  version: 9a65a61ceb776897574c00c289236a95b4b50003
   subpackages:
   - cpu
   - host
@@ -57,7 +57,7 @@ imports:
   subpackages:
   - spew
 - name: github.com/docker/distribution
-  version: cd27f179f2c10c5d300e6d09025b538c475b0d51
+  version: f4118485915abb8b163442717326597908eee6aa
   subpackages:
   - digest
   - reference
