+++ glide.yaml
@@ -6,6 +6,8 @@ import:
     repo: git@github.com:DataDog/datadog-agent.git
   - package: github.com/DataDog/agent-payload
   - package: github.com/DataDog/gopsutil
+    vcs: git
+    version: db/add_windows_checks
     subpackages:
     - cpu
     - host
